requirements.txt
mutagen==1.45.1
Pillow==8.0.1
pydub==0.24.1
pygame==2.0.0
